Login Code
Beginning login code. Got the SQL to retrieve a users cookie ID and user
ID on login once theyve input their username and password. This info
will be used to create a logged in cookie to maintain a session. Ugly
login form added to create an account page for testing purposes for time
being.